#192 fix cassandra integration tests
