+++ .travis.yml
@@ -136,16 +136,16 @@ jobs:
     jdk: oraclejdk11
     install: true
     script: mvn -B -Dtest -P aws-sdk-2 install
-#  - name: "Cassandra [oraclejdk8]"
-#    stage: test
-#    jdk: oraclejdk8
-#    install: true
-#    script: mvn -B -Dtest -P cassandra install
-#  - name: "Cassandra [oraclejdk11]"
-#    stage: test
-#    jdk: oraclejdk11
-#    install: true
-#    script: mvn -B -Dtest -P cassandra install
+  - name: "Cassandra [oraclejdk8]"
+    stage: test
+    jdk: oraclejdk8
+    install: true
+    script: mvn -B -Dtest -P cassandra install
+  - name: "Cassandra [oraclejdk11]"
+    stage: test
+    jdk: oraclejdk11
+    install: true
+    script: mvn -B -Dtest -P cassandra install
 env:
   global:
   - secure: d1B5Y4hSxFBgXPKEkZkBTsCgccZAB72G9A8Rxx/cRH1lmSRqY3GSma37Np9hG9QmxXxvzfPkFvJO7chkvn4HpbQSHOVrqGDkzHmiG7lVfqKMO3vzlR/Gf/2eV8aIhwxA9cPiStBV4glhamcAc+wHUiZPG7ETDbeNEfZ6hLbSaLv1P6yZoOwL/OqMXvFCPZLpG1I1k0bHusGYY35QbKsjpv9ThmCucqn63g+e6OgCGDenxT4w/7lYq+ac2Rj+aJFWeGZk9Dr2jUMa9awhCG3iHEr3atEb6kiWiw9ZvvWErFYkayoSlgGVSvsELQLo3Psnso1UU8p1NwNoSkbFZ46Ojs8MnubE/Fh7AdDMGYxodvXILjqOOuHq7j+pJcwPFSl5xdrw97NTELwZdxmc5nazf8q3jr05jeoWBTCd6hGtRpFFkt5wpOBDXhm/6fNxCRwSnruqn/zn936HSHPUSd9VodYqKxPmEDG7gawo6WIN+vKlmM4kyEFY8J0xEdo5+ePusm8puqlp0mi6bR7X+BPII4zRpL7tj0O3Yw2RoJrZsrw6QjxU4m/bsVty7Eiq5fQV4/v8Wj5DD3OjsFwgwtWQsxTYd82iIMg+se/taO2AB/+SRgZilLUdEKPKOq9RGWGukxKG+12doXkbBPpc1zGteurTjpD9ZR6YSTcK14WfRKQ=
