+++ test/cassandra-3.0.0/src/test/java/io/opentracing/contrib/specialagent/test/cassandra/Cassandra3ITest.java
@@ -16,24 +16,23 @@
 package io.opentracing.contrib.specialagent.test.cassandra;
 
 import java.io.File;
-import java.net.InetSocketAddress;
 
 import org.cassandraunit.utils.EmbeddedCassandraServerHelper;
 
-import com.datastax.oss.driver.api.core.CqlSession;
-import com.datastax.oss.driver.api.core.cql.ResultSet;
-import com.datastax.oss.driver.internal.core.metadata.DefaultEndPoint;
+import com.datastax.driver.core.Cluster;
+import com.datastax.driver.core.ResultSet;
+import com.datastax.driver.core.Session;
 
 import io.opentracing.contrib.specialagent.TestUtil;
 
-public class Cassandra4ITest {
+public class Cassandra3ITest {
   public static void main(final String[] args) throws Exception {
     if (!System.getProperty("java.version").startsWith("1.8.")) {
       System.err.println("Cassandra only works with jdk1.8.");
       System.exit(0);
     }
 
-    System.getProperties().setProperty("java.library.path", new File("src/main/resources/libs").getAbsolutePath());
+    System.getProperties().setProperty("java.library.path", new File("src/test/resources/libs").getAbsolutePath());
 
     final File triggers = new File("target/triggers");
     triggers.mkdirs();
@@ -42,10 +41,13 @@ public class Cassandra4ITest {
     EmbeddedCassandraServerHelper.startEmbeddedCassandra(60_000); // 1 minute
     // EmbeddedCassandraServerHelper.getSession();
 
-    try (final CqlSession session = CqlSession.builder()
-      .addContactEndPoint(new DefaultEndPoint(new InetSocketAddress("127.0.0.1", 9142)))
-      .withLocalDatacenter("datacenter1")
-      .build()) {
+    try (
+      final Cluster cluster = Cluster.builder()
+        .addContactPoints("127.0.0.1")
+        .withPort(9142)
+        .build();
+      final Session session = cluster.connect();
+    ) {
       final ResultSet resultSet = session.execute("SELECT * FROM system.compaction_history");
       System.out.println("Rows: " + resultSet.all().size());
     }
